Rename and refactor model files
Moved and refactored model files to better align with the project structure. Changed package references and modified private model names to follow Go naming conventions.
