+++ internal/websocket/models/params.go
@@ -1,12 +1,12 @@
-package deribit
+package models
 
 import "encoding/json"
 
-var emptyParams = json.RawMessage("{}")
+var EmptyParams = json.RawMessage("{}")
 
 // privateParams is interface for methods require access_token
-type privateParams interface {
-	setToken(token string)
+type PrivateParams interface {
+	SetToken(token string)
 }
 
 // Token is used to embedded in params for private methods
