+++ pkg/models/set_announcement_as_read_params.go
